requirements.txt
flask
flask-script
flask-wtf
flask-sqlalchemy
wtforms-alchemy
python-dateutil
urllib3
lxml
cssselect
